Location

+++ migrations/20230203101558-create-store.js
@@ -16,10 +16,10 @@ module.exports = {
         type: Sequelize.STRING
       },
       latitude: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       longitude: {
-        type: Sequelize.INTEGER
+        type: Sequelize.FLOAT
       },
       address: {
         type: Sequelize.STRING
